ending with collision logic and with editor maybe

+++ Gui.cpp
@@ -143,7 +143,7 @@ gui::DropDownList::DropDownList(float x,float y,float width,float height,sf::Fon
 			sf::Color(70, 70, 70, 200), sf::Color(150, 150, 150, 200), sf::Color(20, 20, 20, 200),
 		sf::Color(255, 255, 255, 200), sf::Color(255, 255, 255, 255), sf::Color(20, 20, 20, 200));
 
-	for (size_t i = 0; i < nrOfElements; i++)
+	for (int i = 0; i < nrOfElements; i++)
 	{
 		this->list.push_back(
 			new gui::Button(x, y+((i+1)*height), width, height,
@@ -262,13 +262,13 @@ gui::TextureSelector::TextureSelector(
 
 	if (this->sheet.getGlobalBounds().width > this->bounds.getGlobalBounds().width)
 	{
-		this->sheet.setTextureRect(sf::IntRect(0, 0, this->bounds.getGlobalBounds().width,this->sheet.getGlobalBounds().height));
+		this->sheet.setTextureRect(sf::IntRect(0, 0, static_cast<int>(this->bounds.getGlobalBounds().width),static_cast<int>(this->sheet.getGlobalBounds().height)));
 	
 	}
 
 	if (this->sheet.getGlobalBounds().height > this->bounds.getGlobalBounds().height)
 	{
-		this->sheet.setTextureRect(sf::IntRect(0, 0, this->sheet.getGlobalBounds().width, this->bounds.getGlobalBounds().height));
+		this->sheet.setTextureRect(sf::IntRect(0, 0, static_cast<int>(this->sheet.getGlobalBounds().width), static_cast<int>(this->bounds.getGlobalBounds().height)));
 	}
 
 	this->selector.setPosition(x +offset, y);
